requirements.txt
langchain-core
langchain-openai
langchain-mcp-adapters
langgraph
python-dotenv
fpdf
uuid
mcp[cli]